Factory method in server actions
