+++ TODO.md
@@ -1,27 +1,34 @@
 ## *Задачи:*
-- ### [x] Работа с БД:
-  - [x] Разобраться с наименованием таблицы;
-  - [x] Пользовательские данные из **json -> db3** и в одной DB таблица с данными, таблица с очередью команд:
-      - [x] При регистрации создание БД пользователя с колонками; 
-      - [x] Перевод из requests в users не корректно; 
-      - [x] Заполнить колонки значениями при регистрации:
-        - [x] Надо как-то разложить файл json на значения и записать их в БД; 
-        - [x] Переделать принимаемый параметр при insert с **std::vector** в  **std::map -> std::pair**;
-  - [x] INSERT в users отправляет **std::vector**, а должен **std::pair**:
-    - [x] Делать ли шаблонным insert ? (НЕТ);
-    
+- ### Работа с БД:
+    - [x] Разобраться с наименованием таблицы;
+    - [x] Пользовательские данные из **json -> db3** и в одной DB таблица с данными, таблица с очередью команд:
+        - [x] При регистрации создание БД пользователя с колонками;
+        - [x] Перевод из requests в users не корректно;
+        - [x] Заполнить колонки значениями при регистрации:
+            - [x] Надо как-то разложить файл json на значения и записать их в БД;
+            - [x] Переделать принимаемый параметр при insert с **std::vector** в  **std::map -> std::pair**;
+    - [x] INSERT в users отправляет **std::vector**, а должен **std::pair**:
+        - [x] Делать ли шаблонным insert ? (НЕТ);
+            - [x] Можно ли как-то еще проверить не пустая БД, кроме как SQLite::Statement ? (не актуально);
+        - [x] Неправильно работает **director.SetBuilder** (не меняет builder-ов) (Изменил указатель на ссылку);
 
-- ### [ ] Действия админа:
-    - [x] Можно ли как-то еще проверить не пустая БД, кроме как SQLite::Statement ? (не актуально);
-    - [x] Неправильно работает **director.SetBuilder** (не меняет builder-ов) (Изменил указатель на ссылку);
-    - [x] Некорректная отправка запроса клиенту;
-    - [ ] Действия при получении данных от клиента (уже зарегистрированного) при повторном подключении
-        - [x] Скорее всего надо переделывать отправку своих данных от клиента серверу.<br>Не получается нормальное внесение данных в БД при MemberInfo;
-        - [ ] Если нет больше запросов к клиенту, то разрываем соединение:
-          - [x] сделать алгоритм действий при отправке запроса отдельным методом;
-          - [ ] как завершить сеанс клиенту ? 
-
-
-- ### [ ] Структура меню:
+- ### Структура меню:
     - [ ] может быть добавить возможность выхода назад в меню user при выборе user и выборе какой зарос отсылать ему ???
     - [ ] при завершении работы админа (turn_off) он не выключается т.к. запущен в другом потоке ??
+    - [x] заменить erase в user на remove erase
+
+- ### Сервер:
+    - [x] Некорректная отправка запроса клиенту;
+    - [x] Action при получении данных от клиента занести в отдельный namespace (Не стал так делать);
+      - [x] Изменить проверку на регистрацию;
+        - [x] создать отдельный класс для проверки зарегистрирован ли пользователь или нет;
+        - [x] что-то не так с логикой, если приходят данные с кейсов DATE;
+    - [x] Все фабрики разбить на отдельные cpp;
+    - [ ] Действия при получении данных от клиента (уже зарегистрированного) при повторном подключении
+        - [x] Скорее всего надо переделывать отправку своих данных от клиента серверу.<br>Не получается нормальное внесение данных в БД при MemberInfo;
+        - [x] Если нет больше запросов к клиенту, то разрываем соединение:
+            - [x] сделать алгоритм действий при отправке запроса отдельным методом;
+            - [x] как завершить сеанс клиенту ?
+            - [x] как завершить сеанс серверу ?
+        - [ ] Id_cmd int
+          - [ ] insert оставить map<std::string, std::string>, а где нужно преобразуем в int
