+++ include/server/server_actions/s_factory.h
@@ -1,0 +1,10 @@
+#pragma once
+
+#include "nlohmann/json.hpp"
+
+
+class Action {
+public:
+    virtual void execute(nlohmann::json& j) = 0;
+    virtual ~Action() = default;
+};
